Fitness that include depot, and terminate if score is less than or equal benchmark score

+++ src/GeneticAlgorithm/GeneticAlgorithm.java
@@ -67,7 +67,7 @@ public class GeneticAlgorithm {
     /**
      * Get best fitness of Population
      */
-    public int getAlphaFitness() {
+    public double getAlphaFitness() {
         return population.getAlphaFitness();
     }
